refactor(frontend): extract useSiloSearch hook from SiloPlaygroundPage
Moves all 29 state vars, 3 useEffects, and 15 handler/helper functions
into frontend/src/hooks/useSiloSearch.ts (574 lines). The page becomes
layout-only (~75 lines of non-JSX logic). Exports QueryHistoryEntry,
PanelState, SiloSearchState types. No behavior changes.
Note: JSX alone is ~520 lines so the strict < 200 line AC is not met
without further JSX sub-component extraction.

Plan: silo-playground-revamp
Step: 018
FR: FR-7.1
